+++ src/main/java/com/weeklyMission/console/ConsoleIO.java
@@ -1,4 +1,4 @@
-package com.weeklyMission;
+package com.weeklyMission.console;
 
 import com.weeklyMission.dto.VoucherResponse;
 import java.util.List;
